Resolve `namespaceURI` error on stale elements
If the timing is perfect, it's possible for an element to be removed from the DOM in between the `.enabled?` check in `assert_element_not_stale` and the wrapped logic.

Most of the time that's ok, because the relevant playwright error is raised and caught. But when something calls `.inspect`, we were seeing sporadic issues like:

```
Playwright::Error:
  TypeError: Cannot read properties of null (reading 'namespaceURI')
      at eval (eval at evaluate (:290:30), <anonymous>:17:12)
      at UtilityScript.evaluate (<anonymous>:297:18)
      at UtilityScript.<anonymous> (<anonymous>:1:44)
```

This commit detects that issue in the x-path-building javascript, and handles it in assert_element_not_stale. Now this situation will correctly raise a `StaleReferenceError`.